+++ eras/allegra/impl/src/Cardano/Ledger/Allegra.hs
@@ -31,7 +31,6 @@ import Cardano.Ledger.Binary (DecCBOR, EncCBOR)
 import Cardano.Ledger.Block (EraBlockHeader)
 import Cardano.Ledger.Shelley.API
 import qualified Cardano.Ledger.Shelley.Rules as Shelley
-import Data.Bifunctor (Bifunctor (first))
 import Data.List.NonEmpty (NonEmpty)
 import GHC.Generics (Generic)
 
@@ -43,9 +42,9 @@ instance ApplyTx AllegraEra where
 
   mkStAnnTx _ _ _ _ = id
 
-  applyTxValidation validationPolicy globals env state stAnnTx =
-    first AllegraApplyTxError $
-      ruleApplyTxValidation @"LEDGER" validationPolicy globals env state stAnnTx
+  internalApplyTxWithValidation = defaultApplyTxWithValidation @"LEDGER" AllegraApplyTxError
+
+  internalReapplyValidatedTx = defaultReapplyValidatedTx @"LEDGER" AllegraApplyTxError
 
 instance ApplyTick AllegraEra