+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Rules.hs
@@ -5,6 +5,7 @@
 module Cardano.Ledger.Alonzo.Rules (
   module Cardano.Ledger.Alonzo.Rules.Bbody,
   module Cardano.Ledger.Alonzo.Rules.Ledger,
+  module Cardano.Ledger.Alonzo.Rules.Ledgers,
   module Cardano.Ledger.Alonzo.Rules.Utxo,
   module Cardano.Ledger.Alonzo.Rules.Utxos,
   module Cardano.Ledger.Alonzo.Rules.Utxow,
@@ -17,7 +18,7 @@ import Cardano.Ledger.Alonzo.Rules.Deleg ()
 import Cardano.Ledger.Alonzo.Rules.Delegs ()
 import Cardano.Ledger.Alonzo.Rules.Delpl ()
 import Cardano.Ledger.Alonzo.Rules.Ledger
-import Cardano.Ledger.Alonzo.Rules.Ledgers ()
+import Cardano.Ledger.Alonzo.Rules.Ledgers
 import Cardano.Ledger.Alonzo.Rules.Pool ()
 import Cardano.Ledger.Alonzo.Rules.Ppup ()
 import Cardano.Ledger.Alonzo.Rules.Utxo