+++ eras/alonzo/impl/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 1.16.0.0
 
+* Remove `ueUtxo` field from `UtxoEnv`
+* Introduce `LEDGERS` and `STS` instance and wire it to `EraRule LEDGERS`
 * Export `alonzoPlutusScriptDecoder` from `Cardano.Ledger.Alonzo.TxWits`
-* Add `generate-cbor` executable
 * Rename the `Rewarding` verb to `Withdrawing` and deprecate the old names:
   - `AlonzoRewarding` -> `AlonzoWithdrawing`
   - `RewardingPurpose` -> `WithdrawingPurpose`